cart invoice update

+++ soup-back-end/DTOs/Invoice/InvoiceDTO.cs
@@ -5,7 +5,7 @@
         public string paymentId { get; set; } = string.Empty;
         public Guid userId { get; set; } = Guid.Empty;
         public DateTime? invoiceDate { get; set; }
-        public float totalPaid { get; set; }
+        public decimal totalPaid { get; set; }
         public int itemCount { get; set; }
     }
 }
